Refactor: useOutsideClick 으로 파일명 변경!
